Tests/LibWeb/Text/expected/navigation/iframe-navigate-javascript-url.txt

Hello from iframe navigated using javascript url
